feat: Apply neubrutalist design system to PricingComparison component
- Replace rounded corners with sharp geometric edges
- Add bold 4px black borders and brutal shadow effects (8px desktop, 4-6px mobile)
- Implement typography system with display text shadows and uppercase styling
- Use Electric Orange (#FF6B35) as primary accent with gradient effects
- Replace react-icons with self-contained SVG components
- Add responsive layouts: table on desktop, stacked cards on mobile
- Implement hover/touch interactions with shadow depth changes
- Fix header alignment issues for consistent layout across varying content
- Add comprehensive Storybook documentation for design system features
- Create AI-focused pricing comparison story example

Also includes cleanup:
- Remove V2 naming from PeopleAlsoAsk component
- Remove unused dubai-web-dev-content.tsx file
