--- src/components/people-also-ask/dubai-web-dev-content.tsx
@@ -1,182 +1,0 @@
-// Dubai Web Development Companies - People Also Ask Content
-// This file contains the specific PAA questions and answers for our blog
-
-import { PeopleAlsoAsk } from './PeopleAlsoAsk';
-
-export const dubaiWebDevPAAItems = [
-  {
-    question: "How much does a website development cost in Dubai?",
-    answer: (
-      <div className="space-y-3">
-        <p>Website development costs in Dubai vary significantly based on project complexity and requirements:</p>
-        
-        <div className="rounded-lg bg-zinc-100 p-4">
-          <ul className="space-y-2">
-            <li className="flex justify-between">
-              <span><strong>Basic Website (5-10 pages):</strong></span>
-              <span className="font-semibold text-emerald-600">$2,000 - $5,000</span>
-            </li>
-            <li className="flex justify-between">
-              <span><strong>E-commerce Platform:</strong></span>
-              <span className="font-semibold text-emerald-600">$5,000 - $15,000</span>
-            </li>
-            <li className="flex justify-between">
-              <span><strong>Custom Web Application:</strong></span>
-              <span className="font-semibold text-emerald-600">$10,000 - $50,000</span>
-            </li>
-            <li className="flex justify-between">
-              <span><strong>Enterprise Solution:</strong></span>
-              <span className="font-semibold text-emerald-600">$25,000 - $100,000+</span>
-            </li>
-          </ul>
-        </div>
-        
-        <p className="text-sm italic">
-          <strong>Note:</strong> Our AI-powered MVP development starts at <span className="font-bold text-emerald-600">$10,000</span> for a complete 4-week sprint, 
-          delivering a production-ready application with 2-3 core features.
-        </p>
-      </div>
-    )
-  },
-  {
-    question: "Is Dubai good for web developers?",
-    answer: (
-      <div className="space-y-3">
-        <p>Dubai offers exceptional opportunities for web developers with its thriving tech ecosystem:</p>
-        
-        <div className="grid gap-3 md:grid-cols-2">
-          <div className="rounded-lg border border-emerald-200 bg-emerald-50 p-3">
-            <h4 className="mb-2 font-semibold text-emerald-800">Market Advantages</h4>
-            <ul className="space-y-1 text-sm">
-              <li>• AED 1.5B government funding (2023)</li>
-              <li>• 370+ funded startups</li>
-              <li>• Tax-free income</li>
-              <li>• Average salary: $60,000-$120,000</li>
-            </ul>
-          </div>
-          
-          <div className="rounded-lg border border-blue-200 bg-blue-50 p-3">
-            <h4 className="mb-2 font-semibold text-blue-800">Tech Ecosystem</h4>
-            <ul className="space-y-1 text-sm">
-              <li>• 27+ accelerators (Hub71, Area 2071)</li>
-              <li>• DIFC FinTech Hive</li>
-              <li>• Government digital initiatives</li>
-              <li>• International business hub</li>
-            </ul>
-          </div>
-        </div>
-        
-        <p className="text-sm">
-          <strong>Growth Opportunity:</strong> AI web development demand growing at <span className="font-bold text-emerald-600">340% YoY</span>, 
-          with zero specialized local competitors.
-        </p>
-      </div>
-    )
-  },
-  {
-    question: "Which company is best for web development?",
-    answer: (
-      <div className="space-y-3">
-        <p>The best web development company depends on your specific needs:</p>
-        
-        <div className="space-y-3">
-          <div className="rounded-lg border-l-4 border-emerald-500 bg-gray-50 p-4">
-            <h4 className="mb-2 font-semibold">For AI-Powered MVPs (Best Value)</h4>
-            <p className="text-sm">Choose companies specializing in rapid AI development with fixed pricing. 
-            Look for 4-week delivery guarantees and production-ready deployments.</p>
-            <p className="mt-2 text-xs text-zinc-600">
-              <strong>Budget:</strong> $10,000 | <strong>Timeline:</strong> 4 weeks
-            </p>
-          </div>
-          
-          <div className="rounded-lg border-l-4 border-blue-500 bg-gray-50 p-4">
-            <h4 className="mb-2 font-semibold">For Enterprise Solutions</h4>
-            <p className="text-sm">Select agencies with proven enterprise experience, security certifications, 
-            and dedicated support teams.</p>
-            <p className="mt-2 text-xs text-zinc-600">
-              <strong>Budget:</strong> $50,000+ | <strong>Timeline:</strong> 3-6 months
-            </p>
-          </div>
-          
-          <div className="rounded-lg border-l-4 border-purple-500 bg-gray-50 p-4">
-            <h4 className="mb-2 font-semibold">For E-commerce Platforms</h4>
-            <p className="text-sm">Look for specialists in Shopify, WooCommerce, or custom solutions with 
-            payment gateway expertise.</p>
-            <p className="mt-2 text-xs text-zinc-600">
-              <strong>Budget:</strong> $15,000-$30,000 | <strong>Timeline:</strong> 6-8 weeks
-            </p>
-          </div>
-        </div>
-      </div>
-    )
-  },
-  {
-    question: "How to choose the right web design company in Dubai?",
-    answer: (
-      <div className="space-y-3">
-        <p>Follow this 5-step selection process to find the perfect web development partner:</p>
-        
-        <ol className="space-y-3">
-          <li className="flex gap-3">
-            <span className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-emerald-500 text-sm font-bold text-white">1</span>
-            <div>
-              <h4 className="font-semibold">Define Your Requirements</h4>
-              <p className="text-sm text-zinc-600">Budget, timeline, features, and technology preferences</p>
-            </div>
-          </li>
-          
-          <li className="flex gap-3">
-            <span className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-emerald-500 text-sm font-bold text-white">2</span>
-            <div>
-              <h4 className="font-semibold">Check Portfolio & Case Studies</h4>
-              <p className="text-sm text-zinc-600">Look for similar projects and industry experience</p>
-            </div>
-          </li>
-          
-          <li className="flex gap-3">
-            <span className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-emerald-500 text-sm font-bold text-white">3</span>
-            <div>
-              <h4 className="font-semibold">Verify Local Presence</h4>
-              <p className="text-sm text-zinc-600">Dubai license, local team, government partnerships</p>
-            </div>
-          </li>
-          
-          <li className="flex gap-3">
-            <span className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-emerald-500 text-sm font-bold text-white">4</span>
-            <div>
-              <h4 className="font-semibold">Compare Pricing Models</h4>
-              <p className="text-sm text-zinc-600">Fixed price vs hourly, payment terms, support included</p>
-            </div>
-          </li>
-          
-          <li className="flex gap-3">
-            <span className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-emerald-500 text-sm font-bold text-white">5</span>
-            <div>
-              <h4 className="font-semibold">Request References</h4>
-              <p className="text-sm text-zinc-600">Contact 2-3 recent clients for feedback</p>
-            </div>
-          </li>
-        </ol>
-        
-        <div className="mt-4 rounded-lg bg-emerald-50 p-3">
-          <p className="text-sm">
-            <strong>Pro Tip:</strong> Look for companies offering fixed-price MVPs with guaranteed delivery dates 
-            to minimize risk and ensure predictable outcomes.
-          </p>
-        </div>
-      </div>
-    )
-  }
-];
-
-// Component with pre-filled Dubai content
-export const DubaiWebDevPAA = () => {
-  return (
-    <PeopleAlsoAsk
-      title="People Also Ask"
-      subtitle="Common questions about web development in Dubai"
-      items={dubaiWebDevPAAItems}
-      defaultOpenIndex={0}
-    />
-  );
-};
